+++ database/models/brazil.py
@@ -17,8 +17,8 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
 
 from enum import StrEnum
-from typing import NamedTuple
 
+# from typing import NamedTuple
 from pydantic import PositiveInt
 from sqlmodel import (
     Field,
@@ -59,6 +59,36 @@ class StateAcronym(StrEnum):
     TO = 'TO'
 
 
+class StateAcronymName(StrEnum):
+    AC = 'Acre'
+    AL = 'Alagoas'
+    AP = 'Amapá'
+    AM = 'Amazonas'
+    BA = 'Bahia'
+    CE = 'Ceará'
+    DF = 'Distrito Federal'
+    ES = 'Espírito Santo'
+    GO = 'Goiás'
+    MA = 'Maranhão'
+    MT = 'Mato Grosso'
+    MS = 'Mato Grosso do Sul'
+    MG = 'Minas Gerais'
+    PA = 'Pará'
+    PB = 'Paraíba'
+    PR = 'Paraná'
+    PE = 'Pernambuco'
+    PI = 'Piauí'
+    RJ = 'Rio de Janeiro'
+    RN = 'Rio Grande do Norte'
+    RS = 'Rio Grande do Sul'
+    RO = 'Rondônia'
+    RR = 'Roraima'
+    SC = 'Santa Catarina'
+    SP = 'São Paulo'
+    SE = 'Sergipe'
+    TO = 'Tocantins'
+
+
 class State(SQLModel, table=True):
     id: int | None = Field(default=None, primary_key=True)
     acronym: StateAcronym = Field(unique=True)
@@ -76,24 +106,24 @@ class City(SQLModel, table=True):
     addresses: list['Address'] = Relationship(back_populates='city')
 
 
-class Coordinates(NamedTuple):
-    latitude: float
-    longitude: float
-    altitude: float
+# class Coordinates(NamedTuple):
+#     latitude: float
+#     longitude: float
+#     altitude: float
 
 
 # Needed until strawberry support auto | None type
 # https://github.com/strawberry-graphql/strawberry/issues/3435
 class AddressBase(SQLModel):
     id: int | None = Field(default=None, primary_key=True)
-    zipcode: PositiveInt | None = None
+    zipcode: int | None = Field(None, gt=1_000_000, lt=99_999_999)
     neighborhood: str | None = None
     complement: str | None = None
     # coordinates: Coordinates | None = None
 
 
 class Address(AddressBase, table=True):
-    zipcode: PositiveInt = Field(unique=True)
+    zipcode: int = Field(unique=True, gt=1_000_000, lt=99_999_999)
 
     state_id: PositiveInt = Field(foreign_key='state.id')
     state: State = Relationship(back_populates='addresses')